# Faster covariance computation for small blocks